LayoutTests/fast/frames/seamless/seamless-quirks.html

<!DOCTYPE html>
<script src="../../js/resources/js-test-pre.js"></script>
<div id="parent" style="width: 200px; height: 200px;">
<iframe id="iframe" seamless src="resources/quirks-square.html"></iframe>
</div>
<script>
debug("Test that a quirks-mode document inside a seamless iframe sizes correctly.")
window.onload = function () {
    window.iframe = document.getElementById("iframe");
    shouldBeTrue("iframe.seamless");
    shouldBeEqualToString("iframe.contentDocument.compatMode", "BackCompat");

    // Quirks mode seamless frames should not expand to fill the window.
    shouldBeEqualToString("window.getComputedStyle(iframe).width", "200px");
    shouldBeEqualToString("window.getComputedStyle(iframe).height", "100px");
}
</script>
