LayoutTests/fast/frames/seamless/seamless-quirks-expected.txt

Test that a quirks-mode document inside a seamless iframe sizes correctly.
FAIL iframe.seamless should be true (of type boolean). Was undefined (of type undefined).
PASS iframe.contentDocument.compatMode is "BackCompat"
FAIL window.getComputedStyle(iframe).width should be 200px. Was 300px.
FAIL window.getComputedStyle(iframe).height should be 100px. Was 150px.
